Encapsulation of legend logic in sankeyDiagramMixin
